extract component `TimersSetHistoryMetric`
